update: fixing long int stuff. we are accessing memories that don't exist. goodbye nostalgia.

+++ Assignment 3/Timer.h
@@ -22,7 +22,7 @@
 #define QUANTUM 300
 
 typedef struct timer {
-	int count;
+	long int count;
 } Timer;
 
 typedef Timer *Timer_p;
